refactor: BetterChat (#3878)
- Refactors better chat to be more readable
- Merges KeepChatAfterDeath into BetterChat
- Makes the chat in the death screen immediately usable (as you'd expect it from the normal chat)
